requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
fuzzywuzzy==0.18.0
matplotlib==3.9.3
networkx==3.3
numpy==1.26.4
peft==0.11.1
Pillow==10.1.0
Pillow==11.0.0
pillow_heif==0.21.0
python-dotenv==1.0.1
pytorch_lightning==2.4.0
Requests==2.32.3
selenium==4.27.1
SQLAlchemy==2.0.31
streamlit==1.36.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.13
timm==0.9.10
torch==2.1.2+cu121
torchvision==0.16.2
tqdm==4.66.4
transformers==4.41.1
webdriver_manager==4.0.2
Werkzeug==3.1.3
sentencepiece==0.2.0
scipy==1.15.2
streamlit-aggrid==1.1.1
